backend/requirements.txt
langchain==0.3.25
langchain_community==0.3.24
langchain_core==0.3.60
langchain_ollama==0.3.3
huggingface_hub
sentence-transformers
chromadb==1.0.10
fastapi==0.115.9
uvicorn==0.34.2
pydantic==2.11.4
python-dotenv==1.1.0
python-multipart
docx2txt==0.9
unstructured==0.17.2
pyngrok 